requirements.txt
numpy
pandas
gdown
tqdm
Pillow
opencv-python-headless
tensorflow
keras
Flask
gunicorn
psycopg2
uwsgi
deepface